Migrate symbols query
